+++ src/core_plugins/kibana/public/dashboard/saved_dashboard/saved_dashboard.js
@@ -1,14 +1,15 @@
 import angular from 'angular';
-import _ from 'lodash';
 import { uiModules } from 'ui/modules';
 import { createDashboardEditUrl } from 'plugins/kibana/dashboard/dashboard_constants';
+import { createLegacyClass } from 'ui/utils/legacy_class';
+
 const module = uiModules.get('app/dashboard');
 
 // Used only by the savedDashboards service, usually no reason to change this
 module.factory('SavedDashboard', function (courier, config) {
   // SavedDashboard constructor. Usually you'd interact with an instance of this.
   // ID is option, without it one will be generated on save.
-  _.class(SavedDashboard).inherits(courier.SavedObject);
+  createLegacyClass(SavedDashboard).inherits(courier.SavedObject);
   function SavedDashboard(id) {
     // Gives our SavedDashboard the properties of a SavedObject
     SavedDashboard.Super.call(this, {
