+++ src/core_plugins/kibana/public/management/sections/objects/_view.js
@@ -9,6 +9,7 @@ import { fatalError, toastNotifications } from 'ui/notify';
 import 'ui/accessibility/kbn_ui_ace_keyboard_mode';
 import { castEsToKbnFieldTypeName } from '../../../../../../utils';
 import { SavedObjectsClientProvider } from 'ui/saved_objects';
+import { isNumeric } from 'ui/utils/numeric';
 
 const location = 'SavedObject view';
 
@@ -55,7 +56,7 @@ uiModules.get('apps/management')
             } catch (err) {
               field.value = field.value;
             }
-          } else if (_.isNumeric(field.value)) {
+          } else if (isNumeric(field.value)) {
             field.type = 'number';
           } else if (Array.isArray(field.value)) {
             field.type = 'array';
