+++ src/core_plugins/kibana/public/visualize/saved_visualizations/_saved_vis.js
@@ -6,18 +6,18 @@
  * NOTE: It's a type of SavedObject, but specific to visualizations.
  */
 
-import _ from 'lodash';
 import { VisProvider } from 'ui/vis';
 import { uiModules } from 'ui/modules';
 import { updateOldState } from 'ui/vis/vis_update_state';
 import { VisualizeConstants } from 'plugins/kibana/visualize/visualize_constants';
+import { createLegacyClass } from 'ui/utils/legacy_class';
 
 uiModules
   .get('app/visualize')
   .factory('SavedVis', function (config, $injector, courier, Promise, savedSearches, Private) {
     const Vis = Private(VisProvider);
 
-    _.class(SavedVis).inherits(courier.SavedObject);
+    createLegacyClass(SavedVis).inherits(courier.SavedObject);
     function SavedVis(opts) {
       const self = this;
       opts = opts || {};
