+++ src/ui/public/agg_response/tabify/_response_writer.js
@@ -3,8 +3,9 @@ import AggConfigResult from 'ui/vis/agg_config_result';
 import { TabifyTable } from 'ui/agg_response/tabify/_table';
 import { TabifyTableGroup } from 'ui/agg_response/tabify/_table_group';
 import { tabifyGetColumns } from 'ui/agg_response/tabify/_get_columns';
+import { createLegacyClass } from '../../utils/legacy_class';
 
-_.class(SplitAcr).inherits(AggConfigResult);
+createLegacyClass(SplitAcr).inherits(AggConfigResult);
 function SplitAcr(agg, parent, key) {
   SplitAcr.Super.call(this, agg, parent, key, key);
 }
