+++ src/ui/public/agg_types/agg_params.js
@@ -1,5 +1,4 @@
 import 'ui/filters/label';
-import _ from 'lodash';
 import { IndexedArray } from 'ui/indexed_array';
 import { FieldParamType } from './param_types/field';
 import { OptionedParamType } from './param_types/optioned';
@@ -7,6 +6,7 @@ import { RegexParamType } from './param_types/regex';
 import { StringParamType } from './param_types/string';
 import { JsonParamType } from './param_types/json';
 import { BaseParamType } from './param_types/base';
+import { createLegacyClass } from '../utils/legacy_class';
 
 const paramTypeMap = {
   field: FieldParamType,
@@ -30,7 +30,7 @@ const paramTypeMap = {
  * @extends IndexedArray
  * @param {object[]} params - array of params that get new-ed up as AggParam objects as descibed above
  */
-_.class(AggParams).inherits(IndexedArray);
+createLegacyClass(AggParams).inherits(IndexedArray);
 function AggParams(params) {
   AggParams.Super.call(this, {
     index: ['name'],
