+++ src/ui/public/directives/rows.js
@@ -4,6 +4,8 @@ import AggConfigResult from 'ui/vis/agg_config_result';
 import { FilterBarClickHandlerProvider } from 'ui/filter_bar/filter_bar_click_handler';
 import { uiModules } from 'ui/modules';
 import tableCellFilterHtml from './partials/table_cell_filter.html';
+import { isNumeric } from '../utils/numeric';
+
 const module = uiModules.get('kibana');
 
 module.directive('kbnRows', function ($compile, $rootScope, getAppState, Private) {
@@ -60,7 +62,7 @@ module.directive('kbnRows', function ($compile, $rootScope, getAppState, Private
 
           // TODO: It would be better to actually check the type of the field, but we don't have
           // access to it here. This may become a problem with the switch to BigNumber
-          if (_.isNumeric(contents)) {
+          if (isNumeric(contents)) {
             $cell.addClass('numeric-value');
           }
         }
