+++ src/ui/public/utils/function.js
@@ -1,0 +1,13 @@
+import _ from 'lodash';
+
+/**
+ * Call all of the function in an array
+ *
+ * @param  {array[functions]} arr
+ * @return {undefined}
+ */
+export function callEach(arr) {
+  return _.map(arr, function (fn) {
+    return _.isFunction(fn) ? fn() : undefined;
+  });
+}
