--- src/ui/public/utils/lodash-mixins/__tests__/index.js
@@ -1,6 +1,0 @@
-import './_move';
-import './_organize_by';
-import './_push_all';
-
-describe('lodash mixins', function () {
-});
