+++ src/ui/public/utils/numeric.js
@@ -1,0 +1,5 @@
+import _ from 'lodash';
+
+export function isNumeric(v) {
+  return !_.isNaN(v) && (typeof v === 'number' || (!Array.isArray(v) && !_.isNaN(parseFloat(v))));
+}
