+++ src/ui/public/utils/simple_emitter.js
@@ -1,5 +1,6 @@
 import _ from 'lodash';
 import { BaseObject } from 'ui/utils/base_object';
+import { createLegacyClass } from './legacy_class';
 
 /**
  * Simple event emitter class used in the vislib. Calls
@@ -7,7 +8,7 @@ import { BaseObject } from 'ui/utils/base_object';
  *
  * @class
  */
-_.class(SimpleEmitter).inherits(BaseObject);
+createLegacyClass(SimpleEmitter).inherits(BaseObject);
 export function SimpleEmitter() {
   this._listeners = {};
 }
